requirements.txt
joblib==1.5.2
langchain==1.2.12
langchain_core==1.2.19
langgraph==1.1.2
lightgbm==4.6.0
matplotlib==3.10.8
numpy==2.4.3
pandas==3.0.1
plotly==6.6.0
python-dotenv==1.2.2
scikit_learn==1.8.0
seaborn==0.13.2
SQLAlchemy==2.0.48
langchain-openai
requests
fastapi
uvicorn
python-multipart